+++ backend/src/main/java/com/github/valfink/backend/mongouser/MongoUser.java
@@ -1,10 +1,10 @@
-package com.github.valfink.backend.user;
+package com.github.valfink.backend.mongouser;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document("users")
-public record User(
+public record MongoUser(
         @Id
         String id,
         String username,
