+++ backend/src/main/java/com/github/valfink/backend/mongouser/MongoUserDTORequest.java
@@ -1,0 +1,7 @@
+package com.github.valfink.backend.mongouser;
+
+public record MongoUserDTORequest(
+        String username,
+        String password
+) {
+}
