+++ backend/src/main/java/com/github/valfink/backend/mongouser/MongoUserDTOResponse.java
@@ -1,0 +1,7 @@
+package com.github.valfink.backend.mongouser;
+
+public record MongoUserDTOResponse(
+        String id,
+        String username
+) {
+}
