--- backend/src/main/java/com/github/valfink/backend/user/UserDTORequest.java
@@ -1,7 +1,0 @@
-package com.github.valfink.backend.user;
-
-public record UserDTORequest(
-        String username,
-        String password
-) {
-}
