--- backend/src/main/java/com/github/valfink/backend/user/UserDTOResponse.java
@@ -1,7 +1,0 @@
-package com.github.valfink.backend.user;
-
-public record UserDTOResponse(
-        String id,
-        String username
-) {
-}
